pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rationale"
version = "0.1.0"
description = "git blame for AI-generated code — code-anchored, repo-local decision log"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Shivam Rawat", email = "shivarawat24@gmail.com"},
]
keywords = ["ai", "claude", "decision-log", "adr", "provenance", "agentic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "anthropic>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[project.scripts]
rationale = "rationale.cli:main"
why = "rationale.cli:why_shortcut"

[project.urls]
Homepage = "https://github.com/shivamrawat/rationale"
Issues = "https://github.com/shivamrawat/rationale/issues"

[tool.setuptools.packages.find]
include = ["rationale", "rationale.*"]
exclude = ["tests*", "rationale.egg-info*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q"
